backend/requirements.txt
autogen==1.0.16
fastapi==0.104.1
pydantic==1.10.12
PyGithub==2.1.1
uvicorn==0.24.0.post1
